Adjust U2F code to make Windows happier.
Several things:
  * Pretend to be a removable device. Otherwise, Windows is looking
    for an MBR.
  * Add the "jmp short; nop" signature at the beginning of the boot
    sector. Otherwise, Windows doesn't recognize this as a valid
    FAT boot sector.
  * Replace \n's with \r\n's.
  * Fix broken FAT date/time handling.

+++ firmware/boot-uf2/uf2.h
@@ -5,7 +5,7 @@
 #include <stdint.h>
 
 struct uf2_configuration {
-  uint16_t total_sectors;
+  uint32_t total_sectors;
   __code const char *info_uf2_txt;
   __code const char *index_htm;
 
@@ -23,7 +23,7 @@ bool uf2_scsi_command (uint8_t lun, __xdata uint8_t *command, uint8_t length) __
 bool uf2_scsi_data_out(uint8_t lun, __xdata const uint8_t *data, uint16_t length) __reentrant;
 bool uf2_scsi_data_in (uint8_t lun, __xdata uint8_t *data, uint16_t length) __reentrant;
 
-bool uf2_fat_read (uint16_t lba, __xdata uint8_t *data);
-bool uf2_fat_write(uint16_t lba, __xdata const uint8_t *data);
+bool uf2_fat_read (uint32_t lba, __xdata uint8_t *data);
+bool uf2_fat_write(uint32_t lba, __xdata const uint8_t *data);
 
 #endif
